+++ packages/desktop/app/AppState.ts
@@ -1,10 +1,7 @@
-import { MessageType } from '../test/TestIpcMessage'
 import { Store } from './javascripts/Main/Store/Store'
 import { StoreKeys } from './javascripts/Main/Store/StoreKeys'
 import { Paths, Urls } from './javascripts/Main/Types/Paths'
 import { UpdateState } from './javascripts/Main/UpdateManager'
-import { handleTestMessage } from './javascripts/Main/Utils/Testing'
-import { isTesting } from './javascripts/Main/Utils/Utils'
 import { WindowState } from './javascripts/Main/Window'
 
 export class AppState {
@@ -27,12 +24,6 @@ export class AppState {
     this.store.set(StoreKeys.LastRunVersion, this.version)
 
     this.updates = new UpdateState(this)
-
-    if (isTesting()) {
-      handleTestMessage(MessageType.AppStateCall, (method, ...args) => {
-        ;(this as any)[method](...args)
-      })
-    }
   }
 
   public isRunningVersionForFirstTime(): boolean {