+++ packages/encryption/src/Domain/StandardException.ts
@@ -1,5 +1,8 @@
 export class StandardException {
-  constructor(public readonly message: string, log = false) {
+  constructor(
+    public readonly message: string,
+    log = false,
+  ) {
     if (log) {
       console.error('StandardException raised: ', message)
     }