+++ packages/api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.35](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.34...@standardnotes/api@1.26.35) (2023-07-28)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.26.34](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.33...@standardnotes/api@1.26.34) (2023-07-27)
 
 **Note:** Version bump only for package @standardnotes/api