+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.56.32](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.31...@standardnotes/mobile@3.56.32) (2023-07-28)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.31](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.30...@standardnotes/mobile@3.56.31) (2023-07-28)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.30](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.29...@standardnotes/mobile@3.56.30) (2023-07-27)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.56.29](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.28...@standardnotes/mobile@3.56.29) (2023-07-27)
 
 **Note:** Version bump only for package @standardnotes/mobile