+++ packages/models/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.46.17](https://github.com/standardnotes/app/compare/@standardnotes/models@1.46.16...@standardnotes/models@1.46.17) (2023-07-28)
+
+**Note:** Version bump only for package @standardnotes/models
+
 ## [1.46.16](https://github.com/standardnotes/app/compare/@standardnotes/models@1.46.15...@standardnotes/models@1.46.16) (2023-07-27)
 
 **Note:** Version bump only for package @standardnotes/models