+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.28.18](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.28.17...@standardnotes/ui-services@1.28.18) (2023-07-28)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
+## [1.28.17](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.28.16...@standardnotes/ui-services@1.28.17) (2023-07-27)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.28.16](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.28.15...@standardnotes/ui-services@1.28.16) (2023-07-27)
 
 **Note:** Version bump only for package @standardnotes/ui-services